requirements.txt
pillow
opencv-python
tesserocr
numpy
dateutils
sanic
mysql-connector